+++ serde_arrow/src/internal/deserialization/date64_deserializer.rs
@@ -1,8 +1,8 @@
 use chrono::{DateTime, Datelike, Utc};
+use marrow::{datatypes::TimeUnit, view::BitsWithOffset};
 use serde::de::Visitor;
 
 use crate::internal::{
-    arrow::{BitsWithOffset, TimeUnit},
     error::{fail, set_default, try_, Context, ContextSupport, Result},
     utils::Mut,
 };
@@ -73,7 +73,7 @@ impl<'a> Date64Deserializer<'a> {
     }
 }
 
-impl<'de> Context for Date64Deserializer<'de> {
+impl Context for Date64Deserializer<'_> {
     fn annotate(&self, annotations: &mut std::collections::BTreeMap<String, String>) {
         set_default(annotations, "field", &self.path);
         set_default(annotations, "data_type", "Date64");