+++ serde_arrow/src/internal/deserialization/fixed_size_list_deserializer.rs
@@ -1,7 +1,7 @@
+use marrow::view::BitsWithOffset;
 use serde::de::{IgnoredAny, SeqAccess, Visitor};
 
 use crate::internal::{
-    arrow::BitsWithOffset,
     error::{fail, set_default, try_, Context, ContextSupport, Error, Result},
     utils::Mut,
 };
@@ -57,7 +57,7 @@ impl<'a> FixedSizeListDeserializer<'a> {
     }
 }
 
-impl<'a> Context for FixedSizeListDeserializer<'a> {
+impl Context for FixedSizeListDeserializer<'_> {
     fn annotate(&self, annotations: &mut std::collections::BTreeMap<String, String>) {
         set_default(annotations, "field", &self.path);
         set_default(annotations, "data_type", "FixedSizeList(..)");