+++ serde_arrow/src/lib.rs
@@ -28,22 +28,21 @@
     doc = r#"
 ## Overview
 
-| Operation     | [`arrow-*`](#features)                                            | [`arrow2-*`](#features)                             |
-|:--------------|:------------------------------------------------------------------|:----------------------------------------------------|
-| Rust to Arrow | [`to_record_batch`], [`to_arrow`]                                 | [`to_arrow2`]                                       |
-| Arrow to Rust | [`from_record_batch`], [`from_arrow`]                             | [`from_arrow2`]                                     |
-| Array Builder | [`ArrayBuilder::from_arrow`]                                      | [`ArrayBuilder::from_arrow2`]                       |
-| Serializer    | [`ArrayBuilder::from_arrow`] + [`Serializer::new`]                | [`ArrayBuilder::from_arrow2`] + [`Serializer::new`] |
-| Deserializer  | [`Deserializer::from_record_batch`], [`Deserializer::from_arrow`] | [`Deserializer::from_arrow2`]                       |
+| Operation     | [`arrow-*`](#features)                                            | [`arrow2-*`](#features)                             | `marrow`                                            |
+|:--------------|:------------------------------------------------------------------|:----------------------------------------------------|:----------------------------------------------------|
+| Rust to Arrow | [`to_record_batch`], [`to_arrow`]                                 | [`to_arrow2`]                                       | [`to_marrow`]                                       |
+| Arrow to Rust | [`from_record_batch`], [`from_arrow`]                             | [`from_arrow2`]                                     | [`from_marrow`]                                     |
+| Array Builder | [`ArrayBuilder::from_arrow`]                                      | [`ArrayBuilder::from_arrow2`]                       | [`ArrayBuilder::from_marrow`]                       |
+| Serializer    | [`ArrayBuilder::from_arrow`] + [`Serializer::new`]                | [`ArrayBuilder::from_arrow2`] + [`Serializer::new`] | [`ArrayBuilder::from_marrow`] + [`Serializer::new`] |
+| Deserializer  | [`Deserializer::from_record_batch`], [`Deserializer::from_arrow`] | [`Deserializer::from_arrow2`]                       | [`Deserializer::from_marrow`]                       |
 "#
 )]
 //!
 //! See also:
 //!
-//! - the [quickstart guide][_impl::docs::quickstart] for more examples of how
-//!   to use this package
-//! - the [status summary][_impl::docs::status] for an overview over the
-//!   supported Arrow and Rust constructs
+//! - the [quickstart guide][_impl::docs::quickstart] for more examples of how to use this package
+//! - the [status summary][_impl::docs::status] for an overview over the supported Arrow and Rust
+//!   constructs
 //!
 //! ## `arrow` Example
 //!
@@ -78,10 +77,11 @@
 //! # fn main() { }
 //! ```
 //!
-//! The `RecordBatch` can then be written to disk, e.g., as parquet using
-//! the [`ArrowWriter`] from the [`parquet`] crate.
+//! The `RecordBatch` can then be written to disk, e.g., as parquet using the [`ArrowWriter`] from
+//! the [`parquet`] crate.
 //!
-//! [`ArrowWriter`]: https://docs.rs/parquet/latest/parquet/arrow/arrow_writer/struct.ArrowWriter.html
+//! [`ArrowWriter`]:
+//!     https://docs.rs/parquet/latest/parquet/arrow/arrow_writer/struct.ArrowWriter.html
 //! [`parquet`]: https://docs.rs/parquet/latest/parquet/
 //!
 //! ## `arrow2` Example
@@ -131,14 +131,19 @@
 //!
 //! # Features:
 //!
-//! The version of `arrow` or `arrow2` used can be selected via features. Per
-//! default no arrow implementation is used. In that case only the base features
-//! of `serde_arrow` are available.
+//! The version of `arrow` or `arrow2` used can be selected via features. Per default no arrow
+//! implementation is used. In that case only the base features of `serde_arrow` are available.
 //!
-//! The `arrow-*` and `arrow2-*` feature groups are compatible with each other.
-//! I.e., it is possible to use `arrow` and `arrow2` together. Within each group
-//! the highest version is selected, if multiple features are activated. E.g,
-//! when selecting  `arrow2-0-16` and `arrow2-0-17`, `arrow2=0.17` will be used.
+//! The `arrow-*` and `arrow2-*` feature groups are compatible with each other. I.e., it is possible
+//! to use `arrow` and `arrow2` together. Within each group the highest version is selected, if
+//! multiple features are activated. E.g, when selecting  `arrow2-0-16` and `arrow2-0-17`,
+//! `arrow2=0.17` will be used.
+//!
+//! Note that because the highest version is selected, the features are not additive. In particular,
+//! it is not possible to use `serde_arrow::to_arrow` for multiple different `arrow` versions at the
+//! same time. Therefore it is not recommended to use the `arrow` and `arrow2` functions directly in
+//! libraries, but rather rely on the [`marrow`] based functionality. The features of `marrow` are
+//! designed to be strictly additive.
 //!
 //! Available features:
 //!
@@ -188,79 +193,20 @@ pub mod _impl {
 
     #[allow(unused)]
     macro_rules! build_arrow_crate {
-        ($arrow_array:ident, $arrow_buffer:ident, $arrow_data:ident, $arrow_schema:ident) => {
+        ($arrow_array:ident, $arrow_schema:ident) => {
             /// A "fake" arrow crate re-exporting the relevant definitions of the
             /// used arrow-* subcrates
             #[doc(hidden)]
             pub mod arrow {
                 /// The raw arrow packages
                 pub mod _raw {
-                    pub use $arrow_array as array;
-                    pub use $arrow_buffer as buffer;
-                    pub use $arrow_data as data;
-                    pub use $arrow_schema as schema;
+                    pub use {$arrow_array as array, $arrow_schema as schema};
                 }
                 pub mod array {
-                    pub use $arrow_array::RecordBatch;
-                    pub use $arrow_array::array::{
-                        Array,
-                        ArrayRef,
-                        ArrowPrimitiveType,
-                        BooleanArray,
-                        DictionaryArray,
-                        FixedSizeBinaryArray,
-                        FixedSizeListArray,
-                        GenericListArray,
-                        GenericBinaryArray,
-                        GenericStringArray,
-                        LargeStringArray,
-                        make_array,
-                        MapArray,
-                        NullArray,
-                        OffsetSizeTrait,
-                        PrimitiveArray,
-                        StringArray,
-                        StructArray,
-                        UnionArray,
-                    };
-                    pub use $arrow_data::ArrayData;
-                }
-                pub mod buffer {
-                    pub use $arrow_buffer::buffer::{Buffer, ScalarBuffer};
+                    pub use $arrow_array::{RecordBatch, array::{Array, ArrayRef}};
                 }
                 pub mod datatypes {
-                    pub use $arrow_array::types::{
-                        ArrowDictionaryKeyType,
-                        ArrowPrimitiveType,
-                        Date32Type,
-                        Date64Type,
-                        Decimal128Type,
-                        DurationMicrosecondType,
-                        DurationMillisecondType,
-                        DurationNanosecondType,
-                        DurationSecondType,
-                        Float16Type,
-                        Float32Type,
-                        Float64Type,
-                        Int16Type,
-                        Int32Type,
-                        Int64Type,
-                        Int8Type,
-                        Time32MillisecondType,
-                        Time32SecondType,
-                        Time64MicrosecondType,
-                        Time64NanosecondType,
-                        TimestampMicrosecondType,
-                        TimestampMillisecondType,
-                        TimestampNanosecondType,
-                        TimestampSecondType,
-                        UInt16Type,
-                        UInt32Type,
-                        UInt64Type,
-                        UInt8Type,
-                    };
-                    pub use $arrow_buffer::ArrowNativeType;
-                    pub use $arrow_schema::{DataType, Field, FieldRef, Schema, TimeUnit, UnionMode};
+                    pub use $arrow_schema::{DataType, Field, FieldRef, Schema, TimeUnit};
                 }
                 pub mod error {
                     pub use $arrow_schema::ArrowError;
@@ -269,25 +215,25 @@ pub mod _impl {
         };
     }
 
-    // arrow-version:insert: #[cfg(has_arrow_{version})] build_arrow_crate!(arrow_array_{version}, arrow_buffer_{version}, arrow_data_{version}, arrow_schema_{version});
-#[cfg(has_arrow_54)] build_arrow_crate!(arrow_array_54, arrow_buffer_54, arrow_data_54, arrow_schema_54);
-#[cfg(has_arrow_53)] build_arrow_crate!(arrow_array_53, arrow_buffer_53, arrow_data_53, arrow_schema_53);
-    #[cfg(has_arrow_52)] build_arrow_crate!(arrow_array_52, arrow_buffer_52, arrow_data_52, arrow_schema_52);
-    #[cfg(has_arrow_51)] build_arrow_crate!(arrow_array_51, arrow_buffer_51, arrow_data_51, arrow_schema_51);
-    #[cfg(has_arrow_50)] build_arrow_crate!(arrow_array_50, arrow_buffer_50, arrow_data_50, arrow_schema_50);
-    #[cfg(has_arrow_49)] build_arrow_crate!(arrow_array_49, arrow_buffer_49, arrow_data_49, arrow_schema_49);
-    #[cfg(has_arrow_48)] build_arrow_crate!(arrow_array_48, arrow_buffer_48, arrow_data_48, arrow_schema_48);
-    #[cfg(has_arrow_47)] build_arrow_crate!(arrow_array_47, arrow_buffer_47, arrow_data_47, arrow_schema_47);
-    #[cfg(has_arrow_46)] build_arrow_crate!(arrow_array_46, arrow_buffer_46, arrow_data_46, arrow_schema_46);
-    #[cfg(has_arrow_45)] build_arrow_crate!(arrow_array_45, arrow_buffer_45, arrow_data_45, arrow_schema_45);
-    #[cfg(has_arrow_44)] build_arrow_crate!(arrow_array_44, arrow_buffer_44, arrow_data_44, arrow_schema_44);
-    #[cfg(has_arrow_43)] build_arrow_crate!(arrow_array_43, arrow_buffer_43, arrow_data_43, arrow_schema_43);
-    #[cfg(has_arrow_42)] build_arrow_crate!(arrow_array_42, arrow_buffer_42, arrow_data_42, arrow_schema_42);
-    #[cfg(has_arrow_41)] build_arrow_crate!(arrow_array_41, arrow_buffer_41, arrow_data_41, arrow_schema_41);
-    #[cfg(has_arrow_40)] build_arrow_crate!(arrow_array_40, arrow_buffer_40, arrow_data_40, arrow_schema_40);
-    #[cfg(has_arrow_39)] build_arrow_crate!(arrow_array_39, arrow_buffer_39, arrow_data_39, arrow_schema_39);
-    #[cfg(has_arrow_38)] build_arrow_crate!(arrow_array_38, arrow_buffer_38, arrow_data_38, arrow_schema_38);
-    #[cfg(has_arrow_37)] build_arrow_crate!(arrow_array_37, arrow_buffer_37, arrow_data_37, arrow_schema_37);
+    // arrow-version:insert:     #[cfg(has_arrow_{version})] build_arrow_crate!(arrow_array_{version}, arrow_schema_{version});
+    #[cfg(has_arrow_54)] build_arrow_crate!(arrow_array_54, arrow_schema_54);
+    #[cfg(has_arrow_53)] build_arrow_crate!(arrow_array_53, arrow_schema_53);
+    #[cfg(has_arrow_52)] build_arrow_crate!(arrow_array_52, arrow_schema_52);
+    #[cfg(has_arrow_51)] build_arrow_crate!(arrow_array_51, arrow_schema_51);
+    #[cfg(has_arrow_50)] build_arrow_crate!(arrow_array_50, arrow_schema_50);
+    #[cfg(has_arrow_49)] build_arrow_crate!(arrow_array_49, arrow_schema_49);
+    #[cfg(has_arrow_48)] build_arrow_crate!(arrow_array_48, arrow_schema_48);
+    #[cfg(has_arrow_47)] build_arrow_crate!(arrow_array_47, arrow_schema_47);
+    #[cfg(has_arrow_46)] build_arrow_crate!(arrow_array_46, arrow_schema_46);
+    #[cfg(has_arrow_45)] build_arrow_crate!(arrow_array_45, arrow_schema_45);
+    #[cfg(has_arrow_44)] build_arrow_crate!(arrow_array_44, arrow_schema_44);
+    #[cfg(has_arrow_43)] build_arrow_crate!(arrow_array_43, arrow_schema_43);
+    #[cfg(has_arrow_42)] build_arrow_crate!(arrow_array_42, arrow_schema_42);
+    #[cfg(has_arrow_41)] build_arrow_crate!(arrow_array_41, arrow_schema_41);
+    #[cfg(has_arrow_40)] build_arrow_crate!(arrow_array_40, arrow_schema_40);
+    #[cfg(has_arrow_39)] build_arrow_crate!(arrow_array_39, arrow_schema_39);
+    #[cfg(has_arrow_38)] build_arrow_crate!(arrow_array_38, arrow_schema_38);
+    #[cfg(has_arrow_37)] build_arrow_crate!(arrow_array_37, arrow_schema_37);
 
     /// Documentation
     pub mod docs {
@@ -326,13 +272,18 @@ pub use crate::internal::array_builder::ArrayBuilder;
 mod arrow_impl;
 
 #[cfg(has_arrow)]
-pub use arrow_impl::api::{from_arrow, from_record_batch, to_arrow, to_record_batch};
+pub use arrow_impl::{from_arrow, from_record_batch, to_arrow, to_record_batch};
 
 #[cfg(has_arrow2)]
 mod arrow2_impl;
 
 #[cfg(has_arrow2)]
-pub use arrow2_impl::api::{from_arrow2, to_arrow2};
+pub use arrow2_impl::{from_arrow2, to_arrow2};
+
+#[deny(missing_docs)]
+mod marrow_impl;
+
+pub use marrow_impl::{from_marrow, to_marrow};
 
 #[deny(missing_docs)]
 /// Helpers that may be useful when using `serde_arrow`
@@ -383,7 +334,6 @@ pub mod utils {
 /// # #[cfg(not(feature="has_arrow2"))]
 /// # fn main() {}
 /// ```
-
 #[deny(missing_docs)]
 pub mod schema {
     pub use crate::internal::schema::{
@@ -399,3 +349,6 @@ pub mod schema {
         };
     }
 }
+
+/// Re-export of the used marrow version
+pub use marrow;