+++ serde_arrow/src/test/deserialization/list.rs
@@ -1,12 +1,15 @@
+use marrow::{
+    datatypes::FieldMeta,
+    view::{ListView, PrimitiveView, View},
+};
+
 use crate::internal::{
-    arrow::{ArrayView, FieldMeta, ListArrayView, PrimitiveArrayView},
-    deserialization::array_deserializer::ArrayDeserializer,
-    testing::assert_error_contains,
+    deserialization::array_deserializer::ArrayDeserializer, testing::assert_error_contains,
 };
 
 #[test]
 fn invalid_offsets() {
-    let reference = ListArrayView {
+    let reference = ListView {
         validity: None,
         offsets: &[],
         meta: FieldMeta {
@@ -14,13 +17,13 @@ fn invalid_offsets() {
             nullable: false,
             metadata: Default::default(),
         },
-        element: Box::new(ArrayView::Int32(PrimitiveArrayView {
+        elements: Box::new(View::Int32(PrimitiveView {
             validity: None,
             values: &[0, 1, 2, 3, 4, 5],
         })),
     };
 
-    let view = ArrayView::List(ListArrayView {
+    let view = View::List(ListView {
         offsets: &[],
         ..reference.clone()
     });
@@ -29,7 +32,7 @@ fn invalid_offsets() {
         "non empty",
     );
 
-    let view = ArrayView::List(ListArrayView {
+    let view = View::List(ListView {
         offsets: &[0, 5, 2],
         ..reference.clone()
     });