+++ Changes.md
@@ -1,5 +1,13 @@
 # Change log
 
+## 0.13.0
+
+- Migrate internal array abstraction to `marrow`
+  - Breaking change: Dictionary data types no longer support sorting
+  - Breaking change: `serde_arrow::Error` no longer implements `From<arrow::error::ArrowError>`
+  - Breaking change: `serde_arrow::Error` no longer implements `From<arrow2::error::Error>`
+- Add APIs to interact with `marorw` arrays directly
+
 ## 0.12.3
 
 - Add `arrow=54` support